+++ paymentsheet-example/src/main/java/com/stripe/android/paymentsheet/example/playground/PaymentSheetPlaygroundActivity.kt
@@ -60,7 +60,6 @@ import com.stripe.android.model.PaymentMethod
 import com.stripe.android.paymentelement.ConfirmCustomPaymentMethodCallback
 import com.stripe.android.paymentelement.EmbeddedPaymentElement
 import com.stripe.android.paymentelement.ExperimentalAnalyticEventCallbackApi
-import com.stripe.android.paymentelement.ExperimentalCustomPaymentMethodsApi
 import com.stripe.android.paymentelement.TapToAddPreview
 import com.stripe.android.paymentelement.WalletButtonsPreview
 import com.stripe.android.paymentelement.rememberEmbeddedPaymentElement
@@ -99,7 +98,6 @@ import kotlinx.coroutines.withContext
 import com.stripe.android.uicore.R as StripeUiCoreR
 
 @OptIn(
-    ExperimentalCustomPaymentMethodsApi::class,
     WalletButtonsPreview::class,
 )
 internal class PaymentSheetPlaygroundActivity :