+++ paymentsheet-example/src/main/java/com/stripe/android/paymentsheet/example/playground/settings/CustomPaymentMethodsSettingsDefinition.kt
@@ -1,11 +1,9 @@
 package com.stripe.android.paymentsheet.example.playground.settings
 
 import com.stripe.android.paymentelement.EmbeddedPaymentElement
-import com.stripe.android.paymentelement.ExperimentalCustomPaymentMethodsApi
 import com.stripe.android.paymentsheet.PaymentSheet
 import com.stripe.android.paymentsheet.example.playground.PlaygroundState
 
-@OptIn(ExperimentalCustomPaymentMethodsApi::class)
 internal object CustomPaymentMethodsSettingDefinition :
     PlaygroundSettingDefinition<CustomPaymentMethodPlaygroundType>,
     PlaygroundSettingDefinition.Saveable<CustomPaymentMethodPlaygroundType> by EnumSaveable(
@@ -76,7 +74,6 @@ enum class CustomPaymentMethodPlaygroundType(
 
 internal const val DEFAULT_CUSTOM_PAYMENT_METHOD_ID = "cpmt_1QpIMNLu5o3P18Zpwln1Sm6I"
 
-@OptIn(ExperimentalCustomPaymentMethodsApi::class)
 private fun createCustomPaymentMethods(disableBillingDetails: Boolean): List<PaymentSheet.CustomPaymentMethod> {
     return listOf(
         PaymentSheet.CustomPaymentMethod(