+++ paymentsheet/src/test/java/com/stripe/android/paymentelement/callbacks/PaymentElementCallbackReferencesTest.kt
@@ -1,7 +1,6 @@
 package com.stripe.android.paymentelement.callbacks
 
 import com.google.common.truth.Truth.assertThat
-import com.stripe.android.paymentelement.ExperimentalCustomPaymentMethodsApi
 import com.stripe.android.utils.PaymentElementCallbackTestRule
 import org.junit.Rule
 import org.junit.Test
@@ -53,7 +52,6 @@ class PaymentElementCallbackReferencesTest {
         assertThat(PaymentElementCallbackReferences[DEFAULT_TEST_KEY]).isNull()
     }
 
-    @OptIn(ExperimentalCustomPaymentMethodsApi::class)
     private fun createCallbacks(): PaymentElementCallbacks {
         return PaymentElementCallbacks.Builder()
             .createIntentCallback { _, _ ->